+++ test/builder/data.dart
@@ -1,5 +1,5 @@
 const testCases = <String, String>{
-'''
+  '''
 [
     {
       "inputs": [
@@ -35,8 +35,7 @@ const testCases = <String, String>{
       "stateMutability": "pure",
       "type": "function"
     }
-]''':
-'''
+]''': '''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -76,7 +75,7 @@ class Retrieve3 {
   final bool var3;
 }
 ''',
-'''
+  '''
 [
     {
       "inputs": [],
@@ -127,8 +126,7 @@ class Retrieve3 {
       "stateMutability": "nonpayable",
       "type": "function"
     }
-]''':
-'''
+]''': '''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -184,7 +182,7 @@ class GiveMeHello {
   final BigInt num2;
 }
 ''',
-'''
+  '''
 [
     {
         "inputs": [],
@@ -199,8 +197,7 @@ class GiveMeHello {
         "stateMutability": "view",
         "type": "function"
     }
-]''':
-'''
+]''': '''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;